CHROMIUM: ext4: Copy symlink data from page
kmapping the page from d_inode->i_mapping like page_getlink() does not
work for ext4 crypto: the data will be put in a scatter gather list and
sent to the encryption engine.
Instead, copy the data to a bounce buffer, to be sure to have a logical
address to put in the sg list.

Relanding of c/453125/

BUG=chromium:698561
TEST=xfstest generic/068 passes. It used to fail on first symlink.
Test on 3.14 veyron and samus
Test on 3.18 cyan
Test that login

Change-Id: Ie69118eea48a40ec3abfd22c93d9ebd1711ca037
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/470026
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
(cherry picked from commit 5432cbe109cb2600fd61a1c4b8cefac97598b700)
Reviewed-on: https://chromium-review.googlesource.com/472187