FROMLIST: mac80211: Fix clang warning about constant operand in logical operation
When clang detects a non-boolean constant in a logical operation it
generates a 'constant-logical-operand' warning. In
ieee80211_try_rate_control_ops_get() the result of strlen(<const str>)
is used in a logical operation, clang resolves the expression to an
(integer) constant at compile time when clang's builtin strlen function
is used.

Change the condition to check for strlen() > 0 to make the constant
operand boolean and thus avoid the warning.

Link: https://patchwork.kernel.org/patch/9668641/
(applied to git.kernel.org jberg/mac80211-next master)

BUG=chromium:702746
TEST=build with clang for edgar

Change-Id: Ib1511f111d3e69586bf5fcecc8232a63d985d20f
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/489183
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>